requirements.txt
Django==2.2.16
djangorestframework==3.10.0
pytz==2020.1
sqlparse==0.4.1
